+++ CHANGELOG.md
@@ -8,6 +8,9 @@
    2. try `python -m morebuiltins.cmd.ui -e re:findall --gui1`
    3. try `python -m morebuiltins.cmd.ui -e re:findall --gui2`
 5. add default doc for `morebuiltins.cmd.ui` --gui
+6. `morebuiltins.functools.FuncSchema` changed
+   1. `parse` will see arg type as `str` while strict=False and no default/annotation
+   2. add `to_string` to FuncSchema
 
 ### 1.1.5 (2024-08-29)
 1. add `utils.get_hash_int`